backend/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
